pyproject.toml
[project]
name = "linkedin-custom-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.6",
    "fastapi>=0.128.0",
    "fastmcp>=2.14.2",
    "httpx>=0.28.1",
    "mcp>=1.25.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
